requirements.txt
pyyaml
numpy
pandas
scipy
netCDF4

# development
pytest